requirements.txt
py-cord==2.6.1
python-dotenv==1.0.1
aiosqlite==0.20.0
tzdata
aiohttp